refactor: introduce: Authenticable
