--- examples/demo/src/controllers/middleware/mod.rs
@@ -1,1 +1,0 @@
-pub mod api_key;
