+++ examples/demo/src/controllers/mod.rs
@@ -1,4 +1,3 @@
 pub mod auth;
-pub mod middleware;
 pub mod notes;
 pub mod user;
